feat: Moving code around
* feat: Moving timeline context to core-ui

* Move state context out

* Setting up standalone context a bit differently

* Allow for this to run in CI

* Upgrade core ui
