+++ src/renderer/KeybindHandler.tsx
@@ -1,9 +1,8 @@
 import React, { useContext } from "react"
 import { GlobalHotKeys, KeyEventName } from "react-hotkeys"
-import { ReactotronContext } from "reactotron-core-ui"
+import { ReactotronContext, StateContext } from "reactotron-core-ui"
 
 import StandaloneContext from "./contexts/Standalone"
-import StateContext from "./contexts/State"
 
 const keyMap = {
   // Tab Navigation
