+++ src/renderer/RootModals.tsx
@@ -1,7 +1,5 @@
 import React, { useContext } from "react"
-import { DispatchActionModal, SubscriptionAddModal, ReactotronContext } from "reactotron-core-ui"
-
-import StateContext from "./contexts/State"
+import { DispatchActionModal, SubscriptionAddModal, ReactotronContext, StateContext } from "reactotron-core-ui"
 
 function RootModals() {
   const {
