+++ src/renderer/util/connectionHelpers.ts
@@ -8,7 +8,7 @@ import {
   FaChrome as IconBrowserChrome,
 } from "react-icons/fa"
 
-import { Connection } from "../contexts/Standalone/manager"
+import { Connection } from "../contexts/Standalone/useStandalone"
 
 const RX_FIREFOX = /Firefox\//
 const RX_SAFARI = /Safari\//
